chores: tests relocation
